Change ids to BigInteger

+++ repository/models/user.py
@@ -1,5 +1,5 @@
 """repository/models/user.py"""
-from sqlalchemy import Column, Integer, String, DateTime, Boolean
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean
 from repository.base import Base
 
 
@@ -7,7 +7,7 @@ class User(Base):
     __tablename__ = "tb_user"
 
     id = Column(Integer(), primary_key=True)
-    twitter_id = Column(Integer(), nullable=False)
+    twitter_id = Column(BigInteger(), nullable=False)
     username = Column(String(), nullable=False)
     created_at = Column(DateTime(), nullable=False)
     description = Column(String())
